+++ frontend/src/Components/Board/Board.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useState, } from 'react'
+import React, {useEffect, useState,} from 'react'
 import style from './Board.module.css'
 import {Link, useNavigate, useParams} from 'react-router-dom'
 import axios from 'axios'
@@ -6,71 +6,8 @@ import {formatDateDayBefore, getButtonColor, getFontColor, getImage, getMainColo
 import Pagination from '../Pagenation/Pagination'
 import BoardTop from '../BoardTop/BoardTop'
 import apiClient from '../../services/apiClient'
-
-const BoardPost = ({post}) => {
-    const navigate = useNavigate()
-
-    const handleMoveToPostMbti = (postMbti) => {
-        navigate(`/board/${postMbti}/`)
-    }
-
-    return (
-        <>
-            <div className={style.board_post}>
-                <div className={style.board_post_left}>
-                    <div className={style.board_post_category}>
-                        <p style={{color: getFontColor(post.post_mbti)}}>{post.category}</p>
-                    </div>
-                    <div className={style.board_post_title}>
-                        <Link to={`/detail/${post.id}?mbti=${post.post_mbti}`}>{post.title}</Link>
-                        <p style={{color: getFontColor(post.post_mbti)}}>[{post.comments}]</p>
-                    </div>
-                    <div className={style.board_post_bottom}>
-                        <div>
-                            <p>{post.author}</p>
-                        </div>
-                        <div>
-                            <p>{formatDateDayBefore(post.created_at)}</p>
-                        </div>
-                        <div className={style.board_like}>
-                            <p>좋아요</p>
-                            <p>{post.likes}</p>
-                        </div>
-                    </div>
-                </div>
-                <div>
-                    <div className={style.board_post_right}>
-                        {post.mbti.map((m) => (
-                            <button
-                                onClick={() => handleMoveToPostMbti(m)}
-                                style={{backgroundColor: getButtonColor(m)}}
-                            >
-                                {m.toUpperCase()}
-                            </button>
-                        ))}
-                    </div>
-                </div>
-            </div>
-            <hr/>
-        </>
-    )
-}
-
-const BoardPostList = ({mbti, posts}) => {
-    useEffect(() => {
-        //post 변경에 다른 값 갱신
-    }, [posts])
-
-    return (
-        <div>
-            {posts.map((post) => (
-                <>
-                    <BoardPost post={post} mbti={mbti}/>
-                </>
-            ))}
-        </div>
-    )
-}
+import BoardPostBox from "./BoardPostBox";
+import useBoardAPI from "../../api/Hooks/useBoardAPI";
 
 const BoardCategory = ({filter, order, categories, onCategoryChanged}) => {
     useEffect(() => {
@@ -136,84 +73,28 @@ const BoardCategory = ({filter, order, categories, onCategoryChanged}) => {
     )
 }
 
-const BoardPostBox = ({mbti, posts}) => {
-    return (
-        <>
-            <BoardPostList mbti={mbti} posts={posts}/>
-        </>
-    )
-}
-
-const BoardSearch = ({onSearch}) => {
-    return (
-        <div className={style.board_search}>
-            <select>
-            </select>
-            <input/>
-            <button>검색</button>
-        </div>
-    )
-}
-
 const Board = () => {
     const {mbti} = useParams()
-    const [posts, setPosts] = useState([])
-    const [categories, setCategories] = useState([])
-    const [totalPage, setTotalPage] = useState(0)
-    const [currentPage, setCurrentPage] = useState(1)
-    const [filter, setFilter] = useState('') //질문, 유머, 창작 등
-    const [order, setOrder] = useState('recent') //recent, like, comment
+    const {
+        loading,
+        error,
+        posts,
+        categories,
+        totalPage,
+        currentPage,
+        order,
+        filter,
+        handleCategoryChanged,
+        handleGetPostListPage,
+    } = useBoardAPI(mbti)
 
-
-    useEffect(() => {
-        handleGetCategory()
-        handlePageChange(currentPage)
-    }, [mbti, currentPage, order, filter])
-
-    const handleCategoryChanged = (type, data) => {
-        if (type === 'filter') {
-            setFilter(data)
-            setCurrentPage(1)
-        } else if (type === 'order') {
-            setOrder(data)
-            setCurrentPage(1)
-        }
+    if (loading) {
+        return <>Loading...</>
     }
 
-    const handleGetCategory = () => {
-        let url = `/api/posts/category/`
-        apiClient
-            .get(url)
-            .then((response) => {
-                setCategories(response.data)
-            })
-            .catch((error) => {
-                console.error('Error during get posts:', error)
-            })
+    if (error) {
+        return <>{error}</>
     }
-
-    const handlePageChange = (page) => {
-        let url = `/api/posts/mbti/${mbti}/?page=${page}`
-        if (filter !== '') {
-            url += `&category=${filter}`
-        }
-        if (order !== '') {
-            url += `&order=${order}`
-        }
-
-        apiClient
-            .get(url)
-            .then((response) => {
-                setPosts(response.data['results'])
-                setTotalPage(response.data['total_page'])
-                setCurrentPage(page)
-            })
-            .catch((error) => {
-                console.error('Error during get posts:', error)
-            })
-    }
-
-    
 
     return (
         <>
@@ -221,9 +102,9 @@ const Board = () => {
                 <BoardTop mbti={mbti}/>
 
                 <div className={style.writeButton}>
-                <Link to={`/write/`}
-                style ={{backgroundColor:getButtonColor(mbti)}}
-                >글쓰기</Link>
+                    <Link to={`/write/`}
+                          style={{backgroundColor: getButtonColor(mbti)}}
+                    >글쓰기</Link>
                 </div>
 
                 <BoardCategory
@@ -234,7 +115,7 @@ const Board = () => {
                 />
                 <BoardPostBox mbti={mbti} posts={posts}/>
 
-                <Pagination currentPage={currentPage} totalPages={totalPage} onPageChange={handlePageChange}/>
+                <Pagination currentPage={currentPage} totalPages={totalPage} onPageChange={handleGetPostListPage}/>
 
                 {/*<BoardSearch />*/}
             </div>
